V-1.5
* Created new file/class "SDKVersion.java".
* This file, allows the app to check the version of andriod. This ability has
been applied to Splash.java and Main.java
* This was created so the app can decided where it can use the action bar
based on app version compatibility. (SDK Version 11 [Honeycomb (3.0)] and up).

* Fixed major problem with Database. (Previously problem only occured with
Gingerbread (2.3))
* Added Time and Area (tables) to Database.
* Created area_rates and time_rates to document the conversion values.

* Added feature to Manifest, that allows Google Play to filter android
devices to correctly be filtered.
(Previously, devices that did not have Mobile Data capability could not
find this App on Google Play).

* Updated GUI (main.xml file), to take allow access to newest Conversion
to be accessed. This is only available for screens that aren't hpdi.
Causes Error on devices with a hdpi or higher screen.
Works fine on ldpi screens.